requirements/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
Django==4.2.7
django-request-logging==0.7.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
psycopg2-binary==2.9.9
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.1.0
wsproto==1.2.0
